* Index parameters in list bricks are now integers * Function contains display errors fixed
Conflicts:
	catroid/res/values-de/strings.xml

+++ catroid/src/org/catrobat/catroid/content/bricks/DeleteItemOfUserListBrick.java
@@ -69,7 +69,7 @@ public class DeleteItemOfUserListBrick extends FormulaBrick implements OnClickLi
 		this.userList = userList;
 	}
 
-	public DeleteItemOfUserListBrick(double value) {
+	public DeleteItemOfUserListBrick(Integer value) {
 		initializeBrickFields(new Formula(value));
 	}
